ci: enable ecore_con IPv6 tests
We have them behind an env variable as this is system specific. Enable
them on CI to fully test these paths as well.
While this works out of the box for the native build on Travis it fails
for our docker builds, as the dockerd does not have ipv6 enabled by
default. Only enable for Codecov for now.

Reviewed-by: Mike Blumenkrantz <michael.blumenkrantz@gmail.com>
Differential Revision: https://phab.enlightenment.org/D11608