doxygen docs: Add missing Elementary group
Elementary has a main page (elementary_main) like other libs (Ecore, Evas...)
but it was missing a doxygen group like the other libs. This group is not
terribly useful but it is manually referenced from dozens of places, resulting
in lots of warnings.
I thought this solution is simpler than removing all the manual references.