build - remove useless header checks for uv.h
we check for uv.h - but never do anything with it. this is configure
noise that just confuses people into thinking we might do something.
we still have dead libuv support but it's enabled with HAVE_LIBUV and
we'd need to link to libuv too... which we don't do and we don't enable
this. removing libuv supporrt entirely is a totally different question
to raise. this here is just removing noise.