eldbus: fix the data size issue.
dbus message requires the void* value for the message data,
the returned value of dbus_message_iter_get_basic() can be mismatched by to the architure,
This could bring the returned va_argues memory corrupted among the series of values.

We need to use the defined type for the compatibility.

@fix