build: decrease the minimal needed meson verion to 0.50
When we jumped from 0.47 to 0.52 to fix problems some users have been
seeing when building we might jumped a bit to far. We want a good
balance between stable and fast building but also mature enough to be in
distro releases and widely available.

We will try to settle for a middle-ground with 0.50 here. If that causes
problems for you building please let us know instantly.