eln - theme - flat theme lands
This is where flat is now good enough to get to master, so ... in it
goes. it needs a lot of cleaning. lots of images no longer used in the
tree. needs wortk on colorclasses and what not. needs polishing for
scaling (much better than old default though). probably needs a
once-over yto ensure things have not been missed.

@feat