elm theme - notification break and feat add - needed
this is bad but it's necessary. need to make the htextblock accept
events so links can be put in it. this then also requires it become
editable to support anchors and items which are needed for
notification links and inline images.

since this has to break nayway, embrace the break and add action box
swallow - and action buttons. the action buttons will probably need a
swallow added for icons still - but debating that still.

@fix / break / feat