unblock draw block2 - xwayland implementations like in weston break efl
we don't get synthetic configurenotifies for positiomn which we SHOULD
get from a wm when a wm places us on screen... this broke efl
rendering as it would block rendering until we were told by the wm
"you have been placed here" which should always happen.

this works around this b0rkage in xwayland land.

@fix