elm - icon - policy change. use theme defined std icon FIRST always
if you set a standard named icon - look in the theme first for it as
this should match the look of the theme, THEN look in the selected
icon theme. this fixes blank icons with selecting and icon theme in
many situations and fixes blank cons when "use elm icon theme" is
selected.

@fix