jxl - don't be so noisy with ERR's unless it's really bad
evas will try every loader until one succeeds in the case that it
doesn't know what format it is by extension. this means often enough
it's not an image and loaders will fail... printing out lots of errors
is not good. loader should be quiet except for really bad issues like
running out of memory or other consistency issues - not parsing of
file issues