efl_ui_slider: fix the error when scrolling.
When we use efl_ui_textbox and efl_ui_slider simultaneously, the slider invokes the general widget's scrolling hold and freeze behavior, which calls of the textbox's scrolling functions with an obvious error: sd->scroller = NULL. So switch off that in the slider and don't notice any lack of functionality.

Saved sd->freeze as a flag for future use.

Test: elementary_test -to Efl.Ui.Slider (before and after).