Eina test: remove POSIX path from sanitize test
The POSIX path "/home/mydir/../myfile" can't be tested on Windows.
In MSYS2, /home is replaced with the installation path os MSYS2,
that is something like C:\foo\bar (depending on the value of the
path given with the MSYS2 installer). So this can't be tested with
the value "/home/myfile".